chore: merge scaffolds of examples catalog & view
also, nit: rename wrapper to scope
